refactor(frontend): Remove AppSearch Wrapper
